Prepare internal code for switching IOSArray to accept jint instead of NSUInteger. 	Change on 2014/09/03 by kstanger <kstanger@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=74726070

+++ jre_emul/Classes/IOSArray.h
@@ -49,7 +49,7 @@
 // DEPRECATED: Use length instead.
 - (NSUInteger)count __attribute__((deprecated));
 
-- (NSString *)descriptionOfElementAtIndex:(NSUInteger)index;
+- (NSString *)descriptionOfElementAtIndex:(jint)index;
 
 // Returns the element type of this array.
 - (IOSClass *)elementType;
